refactor: mid-modal refactors, just need a savepoint
