+++ src/components/Modal/README.md
@@ -43,6 +43,8 @@ import { modalApi } from '@square/maker/components/Modal';
 import DemoModal from 'doc/DemoModal.vue';
 
 export default {
+	name: 'Demo',
+
 	inject: {
 		modalApi,
 	},
@@ -63,7 +65,7 @@ _DemoModal.vue_
 	<m-modal>
 		<m-image
 			class="cover-photo"
-			src="https://picsum.photos/600/300"
+			src="https://picsum.photos/800/300"
 		/>
 		First modal
 
@@ -88,6 +90,8 @@ import { MImage } from '@square/maker/components/Image';
 import DemoStackedModal from 'doc/DemoStackedModal.vue';
 
 export default {
+	name: 'DemoModal',
+
 	components: {
 		MModal,
 		MImage,
@@ -127,7 +131,7 @@ _DemoStackedModal.vue_
 	<m-modal>
 		<m-image
 			class="cover-photo"
-			src="https://picsum.photos/300/600"
+			src="https://picsum.photos/600/600"
 		/>
 		Second stacked modal
 
@@ -144,6 +148,8 @@ import { MImage } from '@square/maker/components/Image';
 import { MModal, modalApi } from '@square/maker/components/Modal';
 
 export default {
+	name: 'DemoStackedModal',
+
 	components: {
 		MImage,
 		MModal,
@@ -341,12 +347,419 @@ export default {
 </script>
 ```
 
+## Examples
+
+### Stacked Modals
+
+<!--
+
+```vue
+<template>
+	<div>
+		<demo-stacked />
+		<m-modal-layer />
+	</div>
+</template>
+
+<script>
+import { MModalLayer } from '@square/maker/components/Modal';
+import DemoStacked from 'doc/DemoStacked.vue';
+
+export default {
+	components: {
+		DemoStacked,
+		MModalLayer,
+	},
+
+	mixins: [
+		MModalLayer.apiMixin,
+	],
+};
+</script>
+```
+
+_DemoStacked.vue_
+
+```vue
+<template>
+	<div>
+		<button @click="openModal">
+			Open modal
+		</button>
+	</div>
+</template>
+
+<script>
+import { modalApi } from '@square/maker/components/Modal';
+import DemoStackedFirstModal from 'doc/DemoStackedFirstModal.vue';
+
+export default {
+	inject: {
+		modalApi,
+	},
+
+	methods: {
+		openModal() {
+			this.modalApi.open(() => <DemoStackedFirstModal />);
+		},
+	},
+};
+</script>
+```
+
+_DemoStackedFirstModal.vue_
+
+```vue
+<template>
+	<m-modal>
+		<m-image
+			class="cover-photo"
+			src="https://picsum.photos/800/300"
+		/>
+		First modal
+
+		<br><br>
+
+		Modals can stack too
+		<button @click="openStackedModal">
+			Open another modal
+		</button>
+
+		<br><br>
+
+		<button @click="modalApi.close()">
+			Close
+		</button>
+	</m-modal>
+</template>
+
+<script>
+import { MModal, modalApi } from '@square/maker/components/Modal';
+import { MImage } from '@square/maker/components/Image';
+import DemoStackedSecondModal from 'doc/DemoStackedSecondModal.vue';
+
+export default {
+	components: {
+		MModal,
+		MImage,
+	},
+
+	inject: {
+		modalApi,
+	},
+
+	methods: {
+		openStackedModal() {
+			this.modalApi.open(() => <DemoStackedSecondModal />);
+		},
+	},
+};
+</script>
+
+<style scoped>
+.cover-photo {
+	width: 100%;
+	height: 300px;
+}
+
+@media screen and (min-width: 1200px) {
+	.cover-photo {
+		width: inherit;
+		height: inherit;
+	}
+}
+</style>
+```
+
+_DemoStackedSecondModal.vue_
+
+```vue
+<template>
+	<m-modal>
+		<m-image
+			class="cover-photo"
+			src="https://picsum.photos/600/600"
+		/>
+		Second stacked modal
+
+		<br><br>
+
+		<button @click="modalApi.close()">
+			Close
+		</button>
+	</m-modal>
+</template>
+
+<script>
+import { MImage } from '@square/maker/components/Image';
+import { MModal, modalApi } from '@square/maker/components/Modal';
+
+export default {
+	components: {
+		MImage,
+		MModal,
+	},
+
+	inject: {
+		modalApi,
+	},
+};
+</script>
+
+<style scoped>
+.cover-photo {
+	width: 100%;
+	height: 300px;
+}
+@media screen and (min-width: 1200px) {
+	.cover-photo {
+		width: inherit;
+		height: inherit;
+	}
+}
+</style>
+```
+
+-->
+
+### Gallery Lightbox
+
+```vue
+<template>
+	<div>
+		<demo-gallery />
+		<m-modal-layer />
+	</div>
+</template>
+
+<script>
+import { MModalLayer } from '@square/maker/components/Modal';
+import DemoGallery from 'doc/DemoGallery.vue';
+
+export default {
+	components: {
+		DemoGallery,
+		MModalLayer,
+	},
+
+	mixins: [
+		MModalLayer.apiMixin,
+	],
+};
+</script>
+```
+
+_DemoGallery.vue_
+
+```vue
+<template>
+	<div class="gallery">
+		<m-image
+			v-for="(src, i) in images"
+			:key="src"
+			:src="src"
+			class="Image"
+			@click="openLightbox(i)"
+		/>
+	</div>
+</template>
+
+<script>
+import { modalApi } from '@square/maker/components/Modal';
+import { MImage } from '@square/maker/components/Image';
+import DemoLightbox from 'doc/DemoLightbox.vue';
+
+export default {
+	components: {
+		MImage,
+		DemoLightbox,
+	},
+
+	inject: {
+		modalApi,
+	},
+
+	data() {
+		return {
+			images: [
+				'https://i.picsum.photos/id/1033/2048/1365.jpg?hmac=zEuPfX7t6U866nzXjWF41bf-uxkKOnf1dDrHXmhcK-Q',
+				'https://i.picsum.photos/id/1047/3264/2448.jpg?hmac=ksy0K4uGgm79hAV7-KvsfHY2ZuPA0Oq1Kii9hqkOCfU',
+				'https://i.picsum.photos/id/1063/4867/3215.jpg?hmac=-ksdmOruOUma2z6ENQo9Yqp9T7lsnokLo8SFfAt-UNU'
+			],
+		};
+	},
+
+	methods: {
+		openLightbox(srcIndex) {
+			this.modalApi.open(h => h(
+				DemoLightbox,
+				{
+					props: {
+						startIndex: srcIndex,
+						images: this.images,
+					},
+				},
+			));
+		},
+	},
+};
+</script>
+
+<style scoped>
+.Image {
+	width: 200px;
+	height: 200px;
+	display: inline-block;
+	cursor: pointer;
+}
+.gallery {
+	position: relative;
+	z-index: 0;
+}
+</style>
+```
+
+_DemoLightbox.vue_
+
+```vue
+<template>
+	<m-modal-container class="container">
+		<div
+			style="display: inline-block;"
+			class="close-container"
+			@click="close"
+		>
+			<x-icon
+				class="close icon"
+			/>
+		</div>
+		<div class="navViewBox">
+			<div
+				style="display: inline-block;"
+				@click="prevSrc"
+			>
+				<chevron-left-icon
+					class="icon"
+				/>
+			</div>
+			<m-image
+				class="image"
+				:src="currentSrc"
+				@click="nextSrc"
+			/>
+			<div
+				style="display: inline-block;"
+				@click="nextSrc"
+			>
+				<chevron-right-icon
+					class="icon"
+				/>
+			</div>
+		</div>
+	</m-modal-container>
+</template>
+
+<script>
+import { MModalContainer, modalApi } from '@square/maker/components/Modal';
+import XIcon from '@square/maker-icons/X';
+import ChevronLeftIcon from '@square/maker-icons/ChevronLeft';
+import ChevronRightIcon from '@square/maker-icons/ChevronRight';
+import { MImage } from '@square/maker/components/Image';
+
+export default {
+	components: {
+		MModalContainer,
+		MImage,
+		XIcon,
+		ChevronLeftIcon,
+		ChevronRightIcon,
+	},
+
+	props: {
+		startIndex: {
+			type: Number,
+			required: true,
+		},
+		images: {
+			type: Array,
+			required: true,
+		},
+	},
+
+	data() {
+		return {
+			currentIndex: this.startIndex,
+		};
+	},
+
+	computed: {
+		currentSrc() {
+			return this.images[this.currentIndex];
+		},
+	},
+
+	methods: {
+		nextSrc() {
+			this.currentIndex = (this.currentIndex + 1) % this.images.length;
+		},
+		prevSrc() {
+			this.currentIndex = ((this.currentIndex - 1) + this.images.length ) % this.images.length;
+		},
+		close() {
+			this.modalApi.close();
+		},
+	},
+
+	inject: {
+		modalApi,
+	},
+};
+</script>
+
+<style scoped>
+.container {
+	padding: 72px;
+	user-select: none;
+}
+.icon {
+	color: white;
+	width: 48px;
+	height: 48px;
+	cursor: pointer;
+}
+.icon.close {
+	width: 24px;
+	height: 24px;
+}
+.close-container {
+	position: absolute;
+	top: calc(72px - 24px - 8px);
+	right: calc(72px + 24px - 8px);
+}
+.navViewBox {
+	display: flex;
+	align-items: center;
+}
+.image {
+	cursor: pointer;
+	height: calc(100vh - 72px * 2);
+}
+</style>
+```
+
 <!-- api-tables:start -->
 ## Modal Slots
 
 | Slot    | Description   |
 | ------- | ------------- |
-| default | modal content |
+| default | Modal content |
+
+
+## ModalContainer Slots
+
+| Slot    | Description             |
+| ------- | ----------------------- |
+| default | modal container content |
