+++ src/components/Modal/index.js
@@ -1,3 +1,4 @@
 export { default as modalApi } from './src/modal-api';
 export { default as MModal } from './src/Modal.vue';
+export { default as MModalContainer } from './src/ModalContainer.vue';
 export { default as MModalLayer } from './src/ModalLayer.vue';
